Fix for Issue #2 Command crashes because of missing IMediatorMap injection
Added a StarlingCommand class to use in place of the default Robotlegs Command class. The StarlingCommand class provides the same default injections that the default Robotlegs Command class does with the required changes so the contextView is a starling.display.DisplayObjectContainer and the mediatorMap property is a IStarlingMediatorMap injection.

Added examples in the readme for using the new StarlingCommand.

Signed-off-by: Omar Gonzalez <omar@almerblank.com>